LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/table_fill64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds (table_fill64.wast:1) assert_equals: expected false but got true
PASS #3 Test that WebAssembly compilation fails (table_fill64.wast:139)
PASS #4 Test that WebAssembly compilation fails (table_fill64.wast:148)
PASS #5 Test that WebAssembly compilation fails (table_fill64.wast:157)
PASS #6 Test that WebAssembly compilation fails (table_fill64.wast:166)
PASS #7 Test that WebAssembly compilation fails (table_fill64.wast:175)
PASS #8 Test that WebAssembly compilation fails (table_fill64.wast:184)
PASS #9 Test that WebAssembly compilation fails (table_fill64.wast:193)
PASS #10 Test that WebAssembly compilation fails (table_fill64.wast:203)
PASS #11 Test that WebAssembly compilation fails (table_fill64.wast:214)
PASS #12 Reinitialize the default imports
FAIL #13 Test that WebAssembly compilation succeeds (table_fill64.wast:1) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't validate: table.fill expects an i32 offset value, got I64, in function at index 3 at {loc} expected true got false
FAIL #14 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #15 Test that a WebAssembly code returns a specific result (table_fill64.wast:27) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:10:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #16 Test that a WebAssembly code returns a specific result (table_fill64.wast:28) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:13:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #17 Test that a WebAssembly code returns a specific result (table_fill64.wast:29) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:16:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #18 Test that a WebAssembly code returns a specific result (table_fill64.wast:30) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:19:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #19 Test that a WebAssembly code returns a specific result (table_fill64.wast:31) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:22:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #20 Test that a WebAssembly code returns a specific result (table_fill64.wast:33) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:25:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #21 Test that a WebAssembly code returns a specific result (table_fill64.wast:34) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:28:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #22 Test that a WebAssembly code returns a specific result (table_fill64.wast:35) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:31:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #23 Test that a WebAssembly code returns a specific result (table_fill64.wast:36) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:34:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #24 Test that a WebAssembly code returns a specific result (table_fill64.wast:37) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:37:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #25 Test that a WebAssembly code returns a specific result (table_fill64.wast:38) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:40:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #26 Test that a WebAssembly code returns a specific result (table_fill64.wast:40) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:43:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #27 Test that a WebAssembly code returns a specific result (table_fill64.wast:41) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:46:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #28 Test that a WebAssembly code returns a specific result (table_fill64.wast:42) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:49:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #29 Test that a WebAssembly code returns a specific result (table_fill64.wast:43) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:52:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #30 Test that a WebAssembly code returns a specific result (table_fill64.wast:44) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:55:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #31 Test that a WebAssembly code returns a specific result (table_fill64.wast:46) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:58:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #32 Test that a WebAssembly code returns a specific result (table_fill64.wast:47) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:61:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #33 Test that a WebAssembly code returns a specific result (table_fill64.wast:48) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:64:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #34 Test that a WebAssembly code returns a specific result (table_fill64.wast:49) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:67:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #35 Test that a WebAssembly code returns a specific result (table_fill64.wast:51) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:70:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #36 Test that a WebAssembly code returns a specific result (table_fill64.wast:52) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:73:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #37 Test that a WebAssembly code returns a specific result (table_fill64.wast:53) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:76:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #38 Test that a WebAssembly code returns a specific result (table_fill64.wast:54) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:79:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #39 Test that a WebAssembly code returns a specific result (table_fill64.wast:56) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:82:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #40 Test that a WebAssembly code returns a specific result (table_fill64.wast:57) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:85:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #41 Test that a WebAssembly code returns a specific result (table_fill64.wast:58) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:88:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #42 Test that a WebAssembly code returns a specific result (table_fill64.wast:60) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:91:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #43 Test that a WebAssembly code returns a specific result (table_fill64.wast:61) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:94:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #44 Test that a WebAssembly code traps (table_fill64.wast:63) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:97:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #45 Test that a WebAssembly code returns a specific result (table_fill64.wast:67) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:100:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #46 Test that a WebAssembly code returns a specific result (table_fill64.wast:68) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:103:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #47 Test that a WebAssembly code returns a specific result (table_fill64.wast:69) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:106:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #48 Test that a WebAssembly code traps (table_fill64.wast:71) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:109:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #49 Test that a WebAssembly code traps (table_fill64.wast:76) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:112:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #50 Test that a WebAssembly code returns a specific result (table_fill64.wast:83) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:115:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #51 Test that a WebAssembly code returns a specific result (table_fill64.wast:84) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:118:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #52 Test that a WebAssembly code returns a specific result (table_fill64.wast:85) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:121:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #53 Test that a WebAssembly code returns a specific result (table_fill64.wast:86) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:124:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #54 Test that a WebAssembly code returns a specific result (table_fill64.wast:87) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:127:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #55 Test that a WebAssembly code returns a specific result (table_fill64.wast:89) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:130:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #56 Test that a WebAssembly code returns a specific result (table_fill64.wast:90) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:133:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #57 Test that a WebAssembly code returns a specific result (table_fill64.wast:91) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:136:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #58 Test that a WebAssembly code returns a specific result (table_fill64.wast:92) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:139:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #59 Test that a WebAssembly code returns a specific result (table_fill64.wast:93) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:142:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #60 Test that a WebAssembly code returns a specific result (table_fill64.wast:94) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:145:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #61 Test that a WebAssembly code returns a specific result (table_fill64.wast:96) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:148:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #62 Test that a WebAssembly code returns a specific result (table_fill64.wast:97) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:151:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #63 Test that a WebAssembly code returns a specific result (table_fill64.wast:98) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:154:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #64 Test that a WebAssembly code returns a specific result (table_fill64.wast:99) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:157:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #65 Test that a WebAssembly code returns a specific result (table_fill64.wast:100) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:160:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #66 Test that a WebAssembly code returns a specific result (table_fill64.wast:102) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:163:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #67 Test that a WebAssembly code returns a specific result (table_fill64.wast:103) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:166:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #68 Test that a WebAssembly code returns a specific result (table_fill64.wast:104) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:169:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #69 Test that a WebAssembly code returns a specific result (table_fill64.wast:105) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:172:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #70 Test that a WebAssembly code returns a specific result (table_fill64.wast:107) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:175:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #71 Test that a WebAssembly code returns a specific result (table_fill64.wast:108) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:178:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #72 Test that a WebAssembly code returns a specific result (table_fill64.wast:109) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:181:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #73 Test that a WebAssembly code returns a specific result (table_fill64.wast:110) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:184:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #74 Test that a WebAssembly code returns a specific result (table_fill64.wast:112) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:187:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #75 Test that a WebAssembly code returns a specific result (table_fill64.wast:113) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:190:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #76 Test that a WebAssembly code returns a specific result (table_fill64.wast:114) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:193:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #77 Test that a WebAssembly code returns a specific result (table_fill64.wast:116) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:196:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #78 Test that a WebAssembly code returns a specific result (table_fill64.wast:117) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:199:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #79 Test that a WebAssembly code traps (table_fill64.wast:119) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:202:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #80 Test that a WebAssembly code returns a specific result (table_fill64.wast:123) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:205:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #81 Test that a WebAssembly code returns a specific result (table_fill64.wast:124) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:208:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #82 Test that a WebAssembly code returns a specific result (table_fill64.wast:125) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:211:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #83 Test that a WebAssembly code traps (table_fill64.wast:127) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:214:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
FAIL #84 Test that a WebAssembly code traps (table_fill64.wast:132) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_fill64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:217:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_fill64.wast.js:246:3 expected true got false
PASS #85 Test that WebAssembly compilation fails (table_fill64.wast:139)
PASS #86 Test that WebAssembly compilation fails (table_fill64.wast:148)
PASS #87 Test that WebAssembly compilation fails (table_fill64.wast:157)
PASS #88 Test that WebAssembly compilation fails (table_fill64.wast:166)
PASS #89 Test that WebAssembly compilation fails (table_fill64.wast:175)
PASS #90 Test that WebAssembly compilation fails (table_fill64.wast:184)
PASS #91 Test that WebAssembly compilation fails (table_fill64.wast:193)
PASS #92 Test that WebAssembly compilation fails (table_fill64.wast:203)
PASS #93 Test that WebAssembly compilation fails (table_fill64.wast:214)
